+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+- refactor: rename `StepBuilder.padding` to `StepBuilder.controlsPadding`
+- fix: back button background color to be not transparent 
+- docs: fix typo and improved README
+- example: improved example app
+
 # 0.1.0+1
 
 * Update intallation step in README